+++ Mod/NplCadLibrary/services/NplCadEnvironment.lua
@@ -51,47 +51,60 @@ local CSGPolygon = commonlib.gettable("Mod.NplCadLibrary.csg.CSGPolygon");
 local CSGVertex = commonlib.gettable("Mod.NplCadLibrary.csg.CSGVertex");
 local CSG = commonlib.gettable("Mod.NplCadLibrary.csg.CSG");
 
-local math_pi = 3.1415926;
-local function is_string(input)
+
+NplCadEnvironment.pi = 3.1415926;
+
+function NplCadEnvironment.is_string(input)
 	if(input and type(input) == "string")then
 		return true;
 	end
 end
-local function is_table(input)
+function NplCadEnvironment.is_table(input)
 	if(input and type(input) == "table")then
 		return true;
 	end
 end
-local function is_number(input)
+function NplCadEnvironment.is_number(input)
 	if(input and type(input) == "number")then
 		return true;
 	end
 end
-local function is_array(input)
+function NplCadEnvironment.is_array(input)
 	if(input and type(input) == "table" and (#input) > 0)then
 		return true;
 	end
 end
 
-local function is_node(input)
+function NplCadEnvironment.is_node(input)
 	if(input and type(input) == "table" and (input.getTypeName) and input.getTypeName()=="Node")then
 		return true;
 	end
 end
-local function is_shape(input)
-	if(is_node(input) and input.hasTag and input:hasTag("shape")) then
+function NplCadEnvironment.is_shape(input)
+	if(NplCadEnvironment.is_node(input) and input.hasTag and input:hasTag("shape")) then
 		return true;
 	end
 end
-local function is_path(input)
+function NplCadEnvironment.is_path(input)
 	if(input and type(input) == "table" and input.expandToCAG and input.innerToCAG and input.rectangularExtrude) then
 		return true;
 	end
 end
 
--- number pi are exposed
-NplCadEnvironment.PI = 3.1415926;
-NplCadEnvironment.pi = NplCadEnvironment.PI;
+local pi = NplCadEnvironment.pi;
+local is_string = NplCadEnvironment.is_string;
+local is_table = NplCadEnvironment.is_table;
+local is_number = NplCadEnvironment.is_number;
+local is_array = NplCadEnvironment.is_array;
+local is_node = NplCadEnvironment.is_node;
+local is_shape = NplCadEnvironment.is_shape;
+local is_path = NplCadEnvironment.is_path;
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Primitive3d.lua");
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Primitive2d.lua");
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Transformation.lua");
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Boolean.lua");
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Extrusion.lua");
+NPL.load("(gl)Mod/NplCadLibrary/services/interface/Others.lua");
 
 function NplCadEnvironment:new(params)
 	params = params or {}
@@ -251,1322 +264,6 @@ function NplCadEnvironment:build(parent_scene_node,code)
 	end
 end
 --end include----------------------------------------------------------------------------------------------------
-function NplCadEnvironment.union()
-	local self = getfenv(2);
-	self:union__();
-end
-function NplCadEnvironment:union__()
-	local node = self:push__();
-	if(node)then
-		node:setTag("csg_action","union");
-	end
-end
-function NplCadEnvironment.difference()
-	local self = getfenv(2);
-	self:difference__();
-end
-function NplCadEnvironment:difference__()
-	local node = self:push__();
-	if(node)then
-		node:setTag("csg_action","difference");
-	end
-end
-function NplCadEnvironment.intersection()
-	local self = getfenv(2);
-	self:intersection__();
-end
-function NplCadEnvironment:intersection__()
-	local node = self:push__();
-	if(node)then
-		node:setTag("csg_action","intersection");
-	end
-end
-function NplCadEnvironment.cube(options)
-	local self = getfenv(2);
-	return self:cube__(options);
-end
-
-function NplCadEnvironment.read_cube(p)
-	local node = Node.create("");
-	local s = 1;
-	local v = nil;
-	local off = {0,0,0};
-	local round = false;
-	local r = 0;
-	local fn = 8;
-
-	if(is_array(p))then
-		v = p;
-	end
-	if(is_table(p) and p.size and is_array(p.size))then v = p.size; end --{ size: [1,2,3] }
-	if(is_table(p) and p.size and not is_array(p.size))then s = p.size; end --{ size: 1 }
-	if(not is_table(p)) then s = p; end -- (2)
-	if(is_table(p) and p.round == true)then
-		round = true;
-		if(v)then
-			if(is_array(v))then
-				r = (v[1]+v[2]+v[3])/30;
-			else 
-				r = s / 10;
-			end
-		end
-	end
-	if(is_table(p) and p.radius)then
-		round = true;
-		r = p.radius;
-	end
-	if(is_table(p) and p.fn)then
-		fn = p.fn; --applies in case of round: true
-	end
-
-	local x = s;
-	local y = s;
-	local z = s; 
-	if(v and is_array(v))then
-		x = v[1];
-		y = v[2];
-		z = v[3]; 
-	end
-   
-	off = {x/2,y/2,z/2}; -- center: false default
-	local o;
-	if(round)then
-		o = CSGModel:new():init(CSGFactory.roundedCube({radius = {x/2,y/2,z/2}, roundradius = r, resolution = fn}),"roundedCube");
-	else
-		o = CSGModel:new():init(CSGFactory.cube({radius = {x/2,y/2,z/2}}),"cube");
-	end
-	if(is_table(p) and p.center and is_array(p.center))then
-		if(p.center[1])then off[1] = 0; else off[1] = x/2;end
-		if(p.center[2])then off[2] = 0; else off[2] = y/2;end
-		if(p.center[3])then off[3] = 0; else off[3] = z/2;end
-	elseif(is_table(p) and p.center == true)then
-		off = {0,0,0};
-	elseif(is_table(p) and p.center == false)then
-		off = {x/2,y/2,z/2};
-	end
-	node:setDrawable(o);
-	if(off[1] ~= 0 or off[2] ~= 0 or off[3] ~= 0)then
-		node:translate(off[1],off[2],off[3]);
-	end
-	return node;
-end
---[[
-cube(); // openscad like
-cube(1);
-cube({size = 1});
-cube({size = {1,2,3}});
-cube({size = 1, center = true}); // default center:false
-cube({size = 1, center = {true,true,false}}); // individual axis center true or false
---]]
-function NplCadEnvironment:cube__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_cube(options);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.sphere(options)
-	local self = getfenv(2);
-	return self:sphere__(options);
-end
-function NplCadEnvironment.read_sphere(p)
-	local node = Node.create("");
-	
-	local r = 1;
-	local fn = 32;
-	local off = {0,0,0};      
-	local type = 'normal';
-   
-	if(is_table(p) and p.r) then 
-		r = p.r;
-	end
-	if(is_table(p) and p.fn) then
-		fn = p.fn;
-	end
-	if(is_table(p) and p.type) then
-		type = p.type;
-	end
-
-	if(not is_table(p)) then
-		r = p;
-	end
-	off = {0,0,0};       -- center: false (default)
-
-	local o;
-	if(type=='geodesic')then
-		--NOTE:Unimplemented
-		--o = geodesicSphere(p);
-		o = CSGModel:new():init(CSGFactory.sphere({radius = r, resolution = fn}),"sphere");
-	else 
-		o = CSGModel:new():init(CSGFactory.sphere({radius = r, resolution = fn}),"sphere");
-	end
-	if(is_table(p) and p.center and is_table(p.center)) then         -- preparing individual x,y,z center
-		if(p.center[1])then
-			off[1] = 0
-		else
-			off[1] = r
-		end
-		if(p.center[2])then
-			off[2] = 0
-		else
-			off[2] = r
-		end
-		if(p.center[3])then
-			off[3] = 0
-		else
-			off[3] = r
-		end
-	elseif(is_table(p) and p.center==true) then
-		off = {0,0,0};
-	elseif(is_table(p) and p.center==false) then
-		off = {r,r,r};
-	end
-	node:setDrawable(o);
-	if(off[1] ~= 0 or off[2] ~= 0 or off[3] ~= 0)then
-		node:translate(off[1],off[2],off[3]);
-	end
-	return node;
-end
---[[
-	sphere();                          // openscad like
-	sphere(1);
-	sphere({r = 2});                    // Note: center = true is default (unlike other primitives, as OpenSCAD)
-	sphere({r = 2, center = false});     // Note: OpenSCAD doesn't support center for sphere but we do
-	sphere({r = 2, center = {true, true, false}}); // individual axis center 
-	sphere({r = 10, fn = 100 });
---]]
-function NplCadEnvironment:sphere__(options)
-	options = options or {};
-	local parent = self:getNode__();
-	local node = NplCadEnvironment.read_sphere(options);
-	parent:addChild(node);
-	return node;
-end
-
-function NplCadEnvironment.cylinder(options,...)
-	local self = getfenv(2);
-	return self:cylinder__(options);
-end
-
-function NplCadEnvironment.read_cylinder(p,...)
-	local node = Node.create("");
-	local r1 = 1;
-	local r2 = 1;
-	local h = 1;
-	local fn = 32;
-	local round = false; 
-	local a = {...};
-	local off = {0,0,0};
-
-	if(is_table(p) and p.d) then
-		r1 = p.d/2;
-		r2 = r1;
-	end
-	if(is_table(p) and p.r) then
-		r1 = p.r; 
-		r2 = p.r; 
-	end
-	if(is_table(p) and p.h) then
-		h = p.h;
-	end
-	if(is_table(p) and (p.r1 or p.r2)) then
-		r1 = p.r1; 
-		r2 = p.r2; 
-		if(p.h)then
-		h = p.h;
-		end
-	end
-	if(is_table(p) and (p.d1 or p.d2)) then
-		r1 = p.d1/2; 
-		r2 = p.d2/2;
-	end
-    
-	if(is_array(a) and a[1] and is_array(a[1])) then
-		a = a[1]; 
-		r1 = a[1]; 
-		r2 = a[2]; 
-		h = a[3]; 
-		if(#a == 4)then
-		fn = a[4];
-		end
-	end
-	if(is_table(p) and p.fn) then
-		fn = p.fn;
-	end
-	if(is_table(p) and p.round==true) then
-		round = true;
-	end
-	local o;
-
-	if(is_table(p) and (p.from and p.to)) then
-		if(round)then
-			o = CSGModel:new():init(CSGFactory.roundedCylinder({from = p.from, to = p.to, radius = r1, resolution = fn}),"roundedCylinder");
-		else
-			o = CSGModel:new():init(CSGFactory.cylinder({from = p.from,to = p.to,radiusStart = r1,radiusEnd = r2,resolution = fn}),"cylinder");
-		end
-	else
-		if(round)then
-			o = CSGModel:new():init(CSGFactory.roundedCylinder({from = {0,0,0}, to = {0,0,h}, radius = r1, resolution = fn}),"roundedCylinder");
-		else
-			o = CSGModel:new():init(CSGFactory.cylinder({from = {0,0,0}, to = {0,0,h}, radiusStart = r1, radiusEnd = r2, resolution = fn}),"cylinder");
-		end
-		local r;
-		if(r1>r2)then
-			r = r1;
-		else
-			r = r2;
-		end
-
-		if(is_table(p) and p.center and is_table(p.center)) then         -- preparing individual x,y,z center
-			if(p.center[1])then
-				off[1] = 0;
-			else
-				off[1] = r;
-			end
-            if(p.center[2])then
-				off[2] = 0;
-			else
-				off[2] = r;
-			end
-			if(p.center[3])then
-				off[3] = -h/2;
-			else
-				off[3] = 0;
-			end
-			
-		elseif(is_table(p) and p.center==true) then 
-			off = {0,0,-h/2};
-		elseif(is_table(p) and p.center==false) then
-			off = {0,0,0};
-		end
-		if(off[1] ~= 0 or off[2] ~= 0 or off[3] ~= 0)then
-			node:translate(off[1],off[2],off[3]);
-		end
-	end
-	node:setDrawable(o);
-	return node;
-end
---[[
-	cylinder({r = 1, h = 10});                 -- openscad like
-	cylinder({d = 1, h = 10});
-	cylinder({r = 1, h = 10, center = true});   -- default: center = false
-	cylinder({r = 1, h = 10, center = {true, true, false}});  -- individual x,y,z center flags
-	cylinder({r1 = 3, r2 = 0, h = 10});
-	cylinder({d1 = 1, d2 = 0.5, h = 10});
-	cylinder({from = {0,0,0}, to = {0,0,10}, r1 = 1, r2 = 2, fn = 50});
-]]
-function NplCadEnvironment:cylinder__(options,...)
-	options = options or {};
-	local parent = self:getNode__();
-	local node = NplCadEnvironment.read_cylinder(options,...)
-	parent:addChild(node);
-	return node;
-end
---[[
-translate({0,0,10});		--create a new parent node and set translation value 
-translate({0,0,10},obj);	--set translation value with obj          
-translate(0,1,0, obj);		--set translation 
---]]
-function NplCadEnvironment.translate(...)
-	local self = getfenv(2);
-	self:translate__(...);
-end
-function NplCadEnvironment:translate__(p1,p2,p3,p4)
-	local x,y,z, options,obj;
-	if(type(p1) == "table") then
-		options = p1;
-		if #options == 3 then
-			x = options[1];
-			y = options[2];
-			z = options[3];
-		elseif #options == 2 then
-			x = options[1];
-			y = 0;
-			z = options[2];
-		else
-			self:internalLog("translate should have 2 or 3 coords");
-			return;
-		end
-		obj = p2;
-	elseif(type(p1) == "number") then
-		if ( type(p2) == "number" and type(p3) == "number" ) then
-			x = p1;
-			y = tonumber(p2);
-			z = tonumber(p3);
-			obj = p4;
-		elseif (type(p2) == "number" and type(p3) ~= "number") then
-			x = p1;
-			y = 0;
-			z = tonumber(p2);
-			obj = p3;	
-		else 
-			self:internalLog("translate should have 2 or 3 coords");	
-		end
-	else
-		self:internalLog("translate should have a coords array");
-		return;		
-	end
-	if(not obj)then
-		obj = self:push__();
-	end
-	if(obj and obj.setTranslation)then
-		obj:setTranslation(x or 0, y or 0, z or 0);
-	end
-end
---[[
-rotate(2);				--create a new parent node and set rotation value          
-rotate(2,obj);			--set rotation value with obj          
-rotate({1,2,3});		--create a new parent node and set rotation value          
-rotate({1,2,3},obj);	--set rotation value with obj  
---]]
-function NplCadEnvironment.rotate(options,obj)
-	local self = getfenv(2);
-	self:rotate__(options,obj);
-end
-function NplCadEnvironment:rotate__(options,obj)
-	if(not options)then return end
-	local x_angle,y_angle,z_angle;
-	if(is_number(options))then
-		x_angle = options;
-		y_angle = options;
-		z_angle = options;
-	end
-	if(is_array(options))then
-		x_angle = options[1] or 0;
-		y_angle = options[2] or 0;
-		z_angle = options[3] or 0;
-	end
-	if(not obj)then
-		obj = self:push__();
-	end
-	local x = x_angle * math_pi / 180;
-	local y = y_angle * math_pi / 180;
-	local z = z_angle * math_pi / 180;
-
-	if(obj and obj.setRotation)then
-		local q =  Quaternion:new();
-		local yaw = y;
-		local roll = z;
-		local pitch = x;
-		q =  q:FromEulerAngles(yaw,roll,pitch);
-		obj:setRotation(q[1],q[2],q[3],q[4]);
-	end
-end
---[[
-scale(2);			--create a new parent node and set scale value          
-scale(2,obj);		--set scale value with obj          
-scale({1,2,3});		--create a new parent node and set scale value          
-scale({1,2,3},obj); --set scale value with obj          
---]]
-function NplCadEnvironment.scale(options,obj)
-	local self = getfenv(2);
-	self:scale__(options,obj);
-end
-function NplCadEnvironment:scale__(options,obj)
-	if(not options)then return end
-	local x,y,z;
-	if(is_number(options))then
-		x = options;
-		y = options;
-		z = options;
-	elseif(is_array(options))then
-		if #options == 3 then
-			x = options[1] or 1;
-			y = options[2] or 1;
-			z = options[3] or 1;
-		elseif #options == 2 then
-			x = options[1] or 1;
-			y = 1;
-			z = options[2] or 1;
-		else
-			self:internalLog("scale should have 2 or 3 coords");
-			return;
-		end
-	else
-		self:internalLog("scale should have a coords array");
-		return;		
-	end
-	if(not obj)then
-		obj = self:push__();
-	end
-	if(obj and obj.setScale)then
-		obj:setScale(x,y,z);
-	end
-end
---[[
-color({r,g,b});		--create a new parent node and set color value 
-color({r,g,b},obj); --set color value with obj 
-color(color_name);		--create a new parent node and set color value with color name
-color(color_name,obj)		--set color value with obj 
---]]
-function NplCadEnvironment.color(options,obj)
-	local self = getfenv(2);
-	self:color__(options,obj);
-end
-function NplCadEnvironment:color__(options,obj)
-	if(not options)then return end
-	local r,g,b;
-	if(is_string(options))then
-		local v = Color.getValue(options);
-		r = v[1];
-		g = v[2];
-		b = v[3];
-	end
-	if(is_array(options))then
-		r = options[1] or 1;
-		g = options[2] or 1;
-		b = options[3] or 1;
-	end
-	if(not obj)then
-		obj = self:push__();
-	end
-	if(obj and obj.setTag)then
-		obj:setTag("color",{r,g,b});
-	end
-end
-function NplCadEnvironment.loadXml(str)
-	local self = getfenv(2);
-	self:loadXml__(str);
-end
-function NplCadEnvironment:loadXml__(str)
-	local parent = self:getNode__();
-	local node = DomParser.loadStr(str)
-	if(node)then
-		parent:addChild(node);
-	end
-end
-
---[[
-		circle
---]]
-function NplCadEnvironment.circle(options)
-	local self = getfenv(2);
-	return self:circle__(options);
-end
---[[
-	circle();                        -- openscad like
-	circle(1); 
-	circle({d: 2, fn:5});
-	circle({r: 2, fn:5});
-	circle({r: 3, center = true});    -- center: false (default)
-	circle({r: 3, center = {true, true}});    -- individual x,z center flags
-]]
-function NplCadEnvironment:circle__(options)
-	options = options or {};
-	local parent = self:getNode__();
-	local node = NplCadEnvironment.read_circle(options)
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_circle(p)
-	local node = Node.create("");
-    local r = 1;
-    local off;
-    local fn = CSGFactory.defaultResolution2D;
-    if(is_table(p) and p.r)then
-        r = p.r;
-    end
-    if(is_table(p) and p.fn)then
-        fn = p.fn;
-    end
-    if(is_number(p))then
-        r = p;
-    end
-    off = { r, r};
-    if(is_table(p) and p.center == true)then
-        off = { 0, 0 };
-    end
-	local o = CAGModel:new():init(CAGFactory.circle({ center = off, radius = r, resolution = fn}),"circle");
-    node:setDrawable(o);
-	node:setTag("shape","circle");
-	return node;
-end
-function NplCadEnvironment.read_circle2(p,...)
-	local node = Node.create("");
-	local r = 1;
-	local fn = CSGFactory.defaultResolution2D;
-	local a = {...};
-	local off = {0,0,0};
-
-	--x
-	if(not is_table(p)) then
-		r = p;
-	end
-	-- {d=x}
-	if(is_table(p) and p.d) then
-		r = p.d/2;
-	end
-	-- {r=x}
-	if(is_table(p) and p.r) then
-		r = p.r; 
-	end
-	--[r,fn]
-	if(is_array(a) and a[1] and is_array(a[1])) then
-		r = a[1]; 
-		if(#a == 2)then
-		fn = a[2];
-		end
-	end
-	-- {fn=x}
-	if(is_table(p) and p.fn) then
-		fn = p.fn;
-	end
-
-	local o;
-	o = CAGModel:new():init(CAGFactory.circle({radius = r, resolution = fn}),"circle");
-
-	-- {center={}}
-	if(is_table(p) and p.center and is_table(p.center)) then         -- preparing individual x,y,z center
-		if(p.center[1])then
-			off[1] = 0;
-		else
-			off[1] = r;
-		end
-		if(p.center[2])then
-			off[2] = 0;
-		else
-			off[2] = r;
-		end
-	-- {center=true}
-	elseif(is_table(p) and p.center==true) then 
-		off = {0,0,0};
-	-- {center=false}
-	elseif(is_table(p) and p.center==false) then
-		off = {r,0,r};
-	end
-	
-	--if(off[0]||off[1]||off[2]) o = o.translate(off);
-	if(off[1] ~= 0 or off[2] ~= 0 or off[3] ~= 0)then
-		node:translate(off[1],off[2],off[3]);
-	end
-
-	node:setDrawable(o);
-	node:setTag("shape","circle");
-	return node;
-end
---[[
-		square
---]]
-function NplCadEnvironment.square(options)
-	local self = getfenv(2);
-	return self:square__(options);
-end
-
-function NplCadEnvironment.read_square(p)
-	local node = Node.create("");
-    local v = { 1, 1 };
-    local off;
-    if(is_number(p))then
-        v = { p, p };
-    end
-    if(is_array(p))then
-        v = p;
-    end
-    if(is_table(p) and p.size)then
-        if(is_number(p.size))then
-            v[1] = p.size;
-            v[2] = p.size;
-        end
-        if(is_array(p.size))then
-            v = p.size
-        end
-    end
-    off = { v[1] /2, v[2] /2, }
-    if(is_table(p) and p.center == true)then
-        off = { 0, 0 };
-    end
-	local o = CAGModel:new():init(CAGFactory.rectangle({ center = off, radius = { v[1] / 2, v[2] / 2 } }),"square");
-    node:setDrawable(o);
-	node:setTag("shape","square");
-	return node;
-end
-function NplCadEnvironment.read_square2(p)
-	local node = Node.create("");
-	local s = 1;
-	local v = nil;
-	local off = {0,0};
-	local round = false;
-	local r = 0;
-	local fn = 8;
-
-	-- [w,h]
-	if(is_array(p))then
-		v = p;
-	end
-	--{ size: [w.h] }
-	if(is_table(p) and p.size and is_array(p.size))then 
-		v = p.size; 
-	end 
-	--{ size: 1 }
-	if(is_table(p) and p.size and not is_array(p.size))then 
-		s = p.size; 
-	end 
-	-- (2)
-	if(not is_table(p)) then 
-		s = p; 
-	end 
-	if(is_table(p) and p.round == true)then
-		round = true;
-		if(v)then
-			if(is_array(v))then
-				r = (v[1]+v[2])/30;
-			else 
-				r = s / 10;
-			end
-		end
-	end
-	if(is_table(p) and p.radius)then
-		round = true;
-		r = p.radius;
-	end
-	if(is_table(p) and p.fn)then
-		fn = p.fn; --applies in case of round: true
-	end
-
-	local x = s;
-	local y = s; 
-	if(v and is_array(v))then
-		x = v[1];
-		y = v[2]; 
-	end
-   
-	off = {x/2,y/2}; -- center: false default
-	local o;
-	if(round)then
-		--NOTE:Unimplemented
-		--o = CAGModel:new():init(CSGFactory.roundedCube({radius = {x/2,y/2,z/2}, roundradius = r, resolution = fn}),"roundedSquare");
-		o = CAGModel:new():init(CAGFactory.rectangle({radius = {x/2,y/2}}),"square");
-	else
-		o = CAGModel:new():init(CAGFactory.rectangle({radius = {x/2,y/2}}),"square");
-	end
-	if(is_table(p) and p.center and is_array(p.center))then
-		if(p.center[1])then off[1] = 0; else off[1] = x/2;end
-		if(p.center[2])then off[2] = 0; else off[2] = y/2;end
-	elseif(is_table(p) and p.center == true)then
-		off = {0,0};
-	elseif(is_table(p) and p.center == false)then
-		off = {x/2,y/2};
-	end
-	node:setDrawable(o);
-	node:setTag("shape","square");
-
-	if(off[1] ~= 0 or off[2] ~= 0)then
-		node:translate(off[1],off[2],0);
-	end
-	return node;
-end
---[[
-square(); // openscad like
-square(1);
-square({size = 1});
-square({size = {1,2}});
-square({size = 1, center = true}); // default center:false
-square({size = 1, center = {true,true}}); // individual axis center true or false
---]]
-function NplCadEnvironment:square__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_square(options);
-	parent:addChild(node);
-	return node;
-end
---[[
-		polygon
---]]
-function NplCadEnvironment.polygon(options)
-	local self = getfenv(2);
-	return self:polygon__(options);
-end
---[[
-polygon({ {0,0},{3,0},{3,3} });                // openscad like
-polygon({ points = { {0,0},{3,0},{3,3},{0,6} });                    
---]]
-function NplCadEnvironment:polygon__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_polygon(options);
-	parent:addChild(node);
-	return node;
-end
-
-function NplCadEnvironment.read_polygon(p)
-	local node = Node.create("");
-	local points = {};
-    if(p.paths and is_array(p.paths) and is_array(p.paths[1]))then  -- pa(th): [[0,1,2],[2,3,1]] (two paths)
-        for j = 1, #(p.paths) do
-            for i = 1, #(p.paths[j]) do
-                points[i] = p.points[p.paths[j][i]];
-            end
-        end
-    elseif(p.paths and is_array(p.paths))then  -- pa(th): [0,1,2,3,4] (single path)
-        for i = 1, #(p.paths) do
-            points[i] = p.points[p.paths[i]];
-        end
-    else
-        if(is_array(p))then
-            points = p;
-        else
-            points = p.points;
-        end
-    end
-	local o = CAGModel:new():init(CAGFactory.polygon(points),"polygon");
-    node:setDrawable(o);
-	node:setTag("shape","polygon");
-	return node;
-end
-function NplCadEnvironment.read_polygon2(p)
-	local node = Node.create("");
-	local points = {};
-	local off = {0,0};
-
-	-- { {0,0},{3,0},{3,3} }
-	if(is_array(p))then
-		points = p;
-	end
-	--{ points = { {0,0},{3,0},{3,3},{0,6} }
-	if(is_table(p) and p.points)then 
-		points = p.points; 
-	end 
-	local o;
-	o = CAGModel:new():init(CAGFactory.polygon(points),"polygon");
-
-	if(is_table(p) and p.center and is_array(p.center))then
-		if(p.center[1])then off[1] = 0; else off[1] = x/2;end
-		if(p.center[2])then off[2] = 0; else off[2] = z/2;end
-	elseif(is_table(p) and p.center == true)then
-		off = {0,0};
-	elseif(is_table(p) and p.center == false)then
-		off = {x/2,z/2};
-	end
-	node:setDrawable(o);
-	node:setTag("shape","polygon");
-
-	if(off[1] ~= 0 or off[2] ~= 0)then
-		node:translate(off[1],0,off[2]);
-	end
-	return node;
-end
-
---[[
-		path2d
-		-- { {0,0},{3,0},{3,3} }
-		--{ points = { {0,0},{3,0},{3,3},{0,6} }
-		-- { arc = {center={0,0,0},radius=1,startangle=0,endangle= 360,resolution=32,maketangent=false}}
-
-A path can be converted to a CAG in two ways:
-expandToCAG()
-innerToCAG()
---]]
-function NplCadEnvironment.path2d(p)
-	-- { {0,0},{3,0},{3,3} }
-    local path;
-	if(is_array(p)) then
-		path = CAGFactory.path2dFromPoints(p);
-
-	--{ points = { {0,0},{3,0},{3,3},{0,6} }
-	elseif(is_table(p)) then 
-		local closed = false;
-		if(p.closed) then
-			closed = p.closed;
-		end
-		if(p.points) then
-			path = CAGFactory.path2dFromPoints(p.points);
-		elseif(p.arc) then
-			path = CAGFactory.path2dFromArc(p.arc);
-		end
-		path.closed = closed;
-	else
-		path = CAGFactory.path2dFromPoints({});
-	end 
-	return path;
-end
---[[
-innerToCAG( path);
---]]
-function NplCadEnvironment.innerToCAG(path)
-	local self = getfenv(2);
-	return self:innerToCAG__(path);
-end
-function NplCadEnvironment:innerToCAG__(path)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_innerToCAG(path);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_innerToCAG(path)
-	local node = Node.create("");
-	local obj = nil;
-	local o;
-	if(is_path(path)) then
-		obj = path;
-		if(obj.closed == true) then
-			o = CAGModel:new():init(obj:innerToCAG(),"innerToCAG");
-			node:setDrawable(o);
-			node:setTag("shape","innerToCAG");
-		else
-			log("the path which innerToCAG should be CLOSED.");
-		end
-	else
-		log("obj isn't a path,cannot be innerToCAG.");
-	end
-	return node;
-end
-
---[[
-expandToCAG(0.1,path);
-expandToCAG({pathradius = 0.1, fn = 8},path);
--- pathradius default is 0.1
--- fn default is 32
---]]
-function NplCadEnvironment.expandToCAG(options,path)
-	local self = getfenv(2);
-	return self:expandToCAG__(options,path);
-end
-function NplCadEnvironment:expandToCAG__(options,path)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_expandToCAG(options,path);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_expandToCAG(options,path)
-	local node = Node.create("");
-	local obj = nil;
-	local o;
-	local pathradius = 0.1;
-	local fn = 32;
-	if(is_path(path)) then
-		obj = path;
-		if (is_number(options)) then 
-			pathradius = options;
-		elseif (is_table(options)) then
-			if(is_number(options.pathradius)) then
-				pathradius = options.pathradius;
-			end
-			if(is_number(options.fn)) then
-				fn = options.fn;
-			end
-		end
-
-		o = CAGModel:new():init(obj:expandToCAG(pathradius,fn),"expandToCAG");
-		node:setDrawable(o);
-		node:setTag("shape","expandToCAG");
-	else
-		log("obj isn't a path,cannot be expandToCAG.");
-	end
-	return node;
-end
-
---[[
-rectangular_extrude(path)
-rectangular_extrude(0.1,path)
-rectangular_extrude({w = 0.1, h = 0.2, fn = 8},path)
--- width default is 1
--- height default is 1
--- fn default is 32
---]]
-
-function NplCadEnvironment.rectangular_extrude(options,path)
-	local self = getfenv(2);
-	return self:rectangular_extrude__(options,path);
-end
-function NplCadEnvironment:rectangular_extrude__(options,path)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_rectangular_extrude(options,path);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_rectangular_extrude(options,path)
-	local node = Node.create("");
-    local w = 1;
-    local h = 1;
-    local fn = CSGFactory.defaultResolution2D;
-    local closed = false;
-    if(p) then
-        w = p.w or w;
-        h = p.h or h;
-        fn = p.fn or fn;
-        closed = p.closed;
-    end
-	if(is_path(path)) then
-		obj = path;
-		local o = CSGModel:new():init(obj:rectangularExtrude(w,h,fn),"rectangular_extrude");
-		node:setDrawable(o);
-		node:setTag("shape","rectangular_extrude");
-	else
-		log("obj isn't a path,cannot be rectangular_extrude.");
-	end
-	return node;
-end
-
---[[
-extrude(shape)
-extrude({0,10,0},shape)
-extrude({offset = {0,10,0}, twistangle = 360, twiststeps = 100},shape)
--- offset default is {0,0,1}
--- twistangle default is 0
--- twiststeps default is 32
---]]
-
-function NplCadEnvironment.linear_extrude(options,shape)
-	local self = getfenv(2);
-	return self:linear_extrude__(options,shape);
-end
-function NplCadEnvironment:linear_extrude__(options,shape)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_linear_extrude(options,shape);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_linear_extrude(options,shape)
-	local node = nil;
-	local obj = nil;
-	local o;
-	local twistangle = 0;
-	local twiststeps = 32
-	local x,y,z = 0,0,1;
-	if(is_shape(shape)) then
-		node = shape;
-		obj = node:getDrawable().cag_node;
-		if(obj.extrude) then
-			if (is_array(options)) then 
-				if(is_number(options[1])) then
-					x = options[1];
-				end
-				if(is_number(options[2])) then
-					y = options[2];
-				end
-				if(is_number(options[3])) then
-					z = options[3];
-				end
-			elseif (is_table(options)) then
-				if(is_table(options.offset)) then
-					if(is_number(options.offset[1])) then
-						x = options.offset[1];
-					end
-					if(is_number(options.offset[2])) then
-						y = options.offset[2];
-					end
-					if(is_number(options.offset[3])) then
-						z = options.offset[3];
-					end
-				end
-				if(is_number(options.twistangle)) then
-					twistangle = options.twistangle;
-				end
-				if(is_number(options.twiststeps)) then
-					twiststeps = options.twiststeps;
-				end
-			end
-			o = CSGModel:new():init(obj:extrude({offset = {x,y,z},twistangle = twistangle,twiststeps = twiststeps}),"extrude");
-			node:setDrawable(o);
-			node:setTag("extrude","linear_extrude");
-		else
-			log("obj isn't a shape,cannot be linear_extrude.");
-		end
-	end
-	return node;
-end
-
---[[
-rotate_extrude
-arguments: options dict with angle and resolution, both optional
-rotate_extrude(angle,shape)
-rotate_extrude({angle = a, fn = 8},shape)
--- angle default is 360
--- fn default is 32
---]]
-
-function NplCadEnvironment.rotate_extrude(options,shape)
-	local self = getfenv(2);
-	return self:rotate_extrude__(options,shape);
-end
-function NplCadEnvironment:rotate_extrude__(options,shape)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_rotate_extrude(options,shape);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_rotate_extrude(p,o)
-	local node = Node.create("");
-
-    local fn = p.fn or CSGFactory.defaultResolution2D;
-    if(fn<3) then
-        fn = 3;
-    end
-    o = o:getDrawable().cag_node;
-    local offset = p.offset;
-    if(offset)then
-        o:transform(Matrix4.translation(offset));
-    end
-    local ps = {};
-    for i = 1, fn do
-        -- o.{x,y} -> rotate([0,0,i:0..360], obj->{o.x,0,o.y})
-        for j = 1, #(o.sides) do
-            -- has o.sides[j].vertex{0,1}.pos (only x,y)
-            local p = {};
-            local m;
-
-            m = Matrix4.rotationZ(i/fn*360);
-            p[1] = vector3d:new(o.sides[j].vertex0.pos[1],0,o.sides[j].vertex0.pos[2]);
-            p[1] = math3d.MatrixMultiplyVector(nil, m, p[1])
-         
-            p[2] = vector3d:new(o.sides[j].vertex1.pos[1],0,o.sides[j].vertex1.pos[2]);
-            p[2] = math3d.MatrixMultiplyVector(nil, m, p[2])
-         
-            m = Matrix4.rotationZ((i+1)/fn*360);
-            p[3] = vector3d:new(o.sides[j].vertex1.pos[1],0,o.sides[j].vertex1.pos[2]);
-            p[3] = math3d.MatrixMultiplyVector(nil, m, p[3])
-         
-            p[4] = vector3d:new(o.sides[j].vertex0.pos[1],0,o.sides[j].vertex0.pos[2]);
-            p[4] = math3d.MatrixMultiplyVector(nil, m, p[4])
-
-            local p1 = CSGPolygon:new():init({
-            CSGVertex:new():init(p[1]),
-            CSGVertex:new():init(p[2]),
-            CSGVertex:new():init(p[3]),
-            CSGVertex:new():init(p[4]),      -- we make a square polygon (instead of 2 triangles)
-            });
-            table.insert(ps,p1);  
-        end
-    end
-    local model = CSGModel:new():init(CSG.fromPolygons(ps));
-	node:setDrawable(model);
-
-   return node;
-end
-function NplCadEnvironment.read_rotate_extrude2(options,shape)
-	local node = nil;
-	local obj = nil;
-	local o;
-	local angle = 360;
-	local fn = CSGFactory.defaultResolution2D;
-	if(is_shape(shape)) then
-		node = shape;
-		obj = node:getDrawable().cag_node;
-		if(obj.rotateExtrude) then
-			if (is_number(options)) then 
-				angle = options;
-			elseif (is_table(options)) then
-				if(is_number(options.angle)) then
-					angle = options.angle;
-				end
-				if(is_number(options.fn)) then
-					fn = options.fn;
-				end
-			end
-			o = CSGModel:new():init(obj:rotateExtrude({angle = angle,resolution = fn}),"rotate_extrude");
-			node:setDrawable(o);	-- cag should be destroy by this.
-			node:setTag("extrude","rotate_extrude");
-		else
-			log("obj isn't a shape,cannot be rotate_extrude.");
-		end
-	end
-	return node;
-end
 
 
---[[
-polyhedron
-     Parameters:
-       points: points list for this polyhedron
-       faces: trangles list for this polyhedron
---]]
-function NplCadEnvironment.polyhedron(options)
-	local self = getfenv(2);
-	return self:polyhedron__(options);
-end
-function NplCadEnvironment:polyhedron__(options)
-	options = options or {};
-	local parent = self:getNode__();
 
-	local node = NplCadEnvironment.read_polyhedron(options);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_polyhedron(options)
-	local node = Node.create("");
-	local o = CSGModel:new():init(CSGFactory.polyhedron(options));
-	if(o ~= nil) then
-		node:setDrawable(o);
-	end
-	return node;
-end
-
-
---[[
-rectangle();
-rectangle({center = x});
-rectangle({center = {x,y}});
-rectangle({radius = w});
-rectangle({radius = {w,h}});
-rectangle({center = {0,0}, radius = {w,h}});
---]]
-function NplCadEnvironment.rectangle(options)
-	local self = getfenv(2);
-	return self:rectangle__(options);
-end
-function NplCadEnvironment:rectangle__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_rectangle(options);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_rectangle(p)
-	local node = Node.create("");
-	local center = {0,0};
-	local radius = {1,1};
-
-	if(is_table(p))then 
-		if(p.center) then
-			if(is_array(p.center))then
-				center[1] = p.center[1];
-				center[2] = p.center[2]; 
-			elseif(is_number(p.center)) then
-				center[1] = p.center;
-				center[2] = p.center; 				
-			end
-		end
-		if(p.radius)then
-			if(is_array(p.radius))then
-				radius[1] = p.radius[1];
-				radius[2] = p.radius[2]; 
-			elseif(is_number(p.radius)) then
-				radius[1] = p.radius;
-				radius[2] = p.radius; 				
-			end			
-		end
-	end 
-
-	local o = CAGModel:new():init(CAGFactory.rectangle({center = center,  radius = radius}),"rectangle");
-	node:setDrawable(o);
-	node:setTag("shape","rectangle");
-	return node;
-end
-
---[[
-roundedRectangle();
-roundedRectangle({center = x});
-roundedRectangle({center = {x,y}});
-roundedRectangle({radius = w});
-roundedRectangle({radius = {w,h}});
-roundedRectangle({center = {0,0}, radius = {w,h}});
-roundedRectangle({center = {0,0}, radius = {w,h},roundradius = 1,resolution = 32});
---]]
-function NplCadEnvironment.roundedRectangle(options)
-	local self = getfenv(2);
-	return self:roundedRectangle__(options);
-end
-function NplCadEnvironment:roundedRectangle__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_roundedRectangle(options);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_roundedRectangle(p)
-	local node = Node.create("");
-	local center = {0,0};
-	local radius = {1,1};
-	local roundradius  = 0.2;
-	local resolution = 8;
-
-	if(is_table(p))then 
-		if(p.center) then
-			if(is_array(p.center))then
-				center[1] = p.center[1];
-				center[2] = p.center[2]; 
-			elseif(is_number(p.center)) then
-				center[1] = p.center;
-				center[2] = p.center; 				
-			end
-		end
-		if(p.radius)then
-			if(is_array(p.radius))then
-				radius[1] = p.radius[1];
-				radius[2] = p.radius[2]; 
-			elseif(is_number(p.radius)) then
-				radius[1] = p.radius;
-				radius[2] = p.radius; 				
-			end			
-		end
-		if(p.roundradius and is_number(p.roundradius))then
-			roundradius = p.roundradius;
-		end
-		if(p.resolution and is_number(p.resolution))then
-			resolution = p.resolution;
-		end
-	end 
-
-	local o = CAGModel:new():init(CAGFactory.roundedRectangle({center = center,  radius = radius, roundradius = roundradius, resolution = resolution}),"roundedRectangle");
-	node:setDrawable(o);
-	node:setTag("shape","roundedRectangle");
-	return node;
-end
-
---[[
-ellipse();
-ellipse({center = x});
-ellipse({center = {x,y}});
-ellipse({radius = w});
-ellipse({radius = {w,h}});
-ellipse({center = {0,0}, radius = {w,h}});
---]]
-function NplCadEnvironment.ellipse(options)
-	local self = getfenv(2);
-	return self:ellipse__(options);
-end
-function NplCadEnvironment:ellipse__(options)
-	options = options or {};
-	local parent = self:getNode__();
-
-	local node = NplCadEnvironment.read_ellipse(options);
-	parent:addChild(node);
-	return node;
-end
-function NplCadEnvironment.read_ellipse(p)
-	local node = Node.create("");
-	local center = {0,0};
-	local radius = {1,1};
-
-	if(is_table(p))then 
-		if(p.center) then
-			if(is_array(p.center))then
-				center[1] = p.center[1];
-				center[2] = p.center[2]; 
-			elseif(is_number(p.center)) then
-				center[1] = p.center;
-				center[2] = p.center; 				
-			end
-		end
-		if(p.radius)then
-			if(is_array(p.radius))then
-				radius[1] = p.radius[1]/2;
-				radius[2] = p.radius[2]/2; 
-			elseif(is_number(p.radius)) then
-				radius[1] = p.radius/2;
-				radius[2] = p.radius/2; 				
-			end			
-		end
-	end 
-
-	local o = CAGModel:new():init(CAGFactory.ellipse({center = center,  radius = radius}),"ellipse");
-	node:setDrawable(o);
-	node:setTag("shape","ellipse");
-	return node;
-end
