+++ NplCadLibrary.vcxproj
@@ -51,6 +51,12 @@
     <None Include="Mod\NplCadLibrary\drawables\CSGModel.lua" />
     <None Include="Mod\NplCadLibrary\services\CSGBuildContext.lua" />
     <None Include="Mod\NplCadLibrary\services\CSGService.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Boolean.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Extrusion.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Others.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Primitive2d.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Primitive3d.lua" />
+    <None Include="Mod\NplCadLibrary\services\interface\Transformation.lua" />
     <None Include="Mod\NplCadLibrary\services\NplCadEnvironment.lua" />
     <None Include="Mod\NplCadLibrary\testcase\TestConnector.lua" />
     <None Include="Mod\NplCadLibrary\testcase\TestCSG.lua" />
